Cors, imagepage (READ DESC)
Added options for cors configuration. These will need to be updated to include Heroku stuff.
Added like/dislike columns to image db. The issue with image page was largely capitalization of db column names, so before pulling make sure stuff matches.
There was an error with naming a prop key, so I changed it to keyprop. should be changed everywhere